docker compose up -d

+++ docker/kernel/project/ViajesAltairis.Application/Features/Admin/AdministrativeDivisions/Dtos/CreateAdministrativeDivisionRequest.cs
@@ -1,3 +1,3 @@
 namespace ViajesAltairis.Application.Features.Admin.AdministrativeDivisions.Dtos;
 
-public record CreateAdministrativeDivisionRequest(long CountryId, long? ParentId, string Name, long TypeId, byte Level);
+public record CreateAdministrativeDivisionRequest(long CountryId, long? ParentId, string Name, long TypeId, int Level);
