+++ docker/kernel/project/ViajesAltairis.Application/Features/Admin/AdministrativeDivisions/Dtos/UpdateAdministrativeDivisionRequest.cs
@@ -1,3 +1,3 @@
 namespace ViajesAltairis.Application.Features.Admin.AdministrativeDivisions.Dtos;
 
-public record UpdateAdministrativeDivisionRequest(long CountryId, long? ParentId, string Name, long TypeId, byte Level);
+public record UpdateAdministrativeDivisionRequest(long CountryId, long? ParentId, string Name, long TypeId, int Level);
