+++ docker/kernel/project/ViajesAltairis.Application/Features/Admin/HotelProviderRoomTypes/Dtos/CreateHotelProviderRoomTypeRequest.cs
@@ -1,3 +1,3 @@
 namespace ViajesAltairis.Application.Features.Admin.HotelProviderRoomTypes.Dtos;
 
-public record CreateHotelProviderRoomTypeRequest(long HotelProviderId, long RoomTypeId, byte Capacity, int Quantity, decimal PricePerNight, long CurrencyId, long ExchangeRateId);
+public record CreateHotelProviderRoomTypeRequest(long HotelProviderId, long RoomTypeId, int Capacity, int Quantity, decimal PricePerNight, long CurrencyId, long ExchangeRateId);
