+++ docker/kernel/project/ViajesAltairis.Application/Features/Admin/HotelProviderRoomTypes/Validators/UpdateHotelProviderRoomTypeValidator.cs
@@ -10,7 +10,7 @@ public class UpdateHotelProviderRoomTypeValidator : AbstractValidator<UpdateHote
         RuleFor(x => x.Id).GreaterThan(0);
         RuleFor(x => x.HotelProviderId).GreaterThan(0);
         RuleFor(x => x.RoomTypeId).GreaterThan(0);
-        RuleFor(x => x.Capacity).GreaterThan((byte)0);
+        RuleFor(x => x.Capacity).GreaterThan(0);
         RuleFor(x => x.Quantity).GreaterThan(0);
         RuleFor(x => x.PricePerNight).GreaterThan(0);
         RuleFor(x => x.CurrencyId).GreaterThan(0);
