+++ docker/kernel/project/ViajesAltairis.Application/Features/Admin/Hotels/Validators/CreateHotelValidator.cs
@@ -9,7 +9,7 @@ public class CreateHotelValidator : AbstractValidator<CreateHotelCommand>
     {
         RuleFor(x => x.CityId).GreaterThan(0);
         RuleFor(x => x.Name).NotEmpty().MaximumLength(200);
-        RuleFor(x => x.Stars).InclusiveBetween((byte)1, (byte)5);
+        RuleFor(x => x.Stars).InclusiveBetween(1, 5);
         RuleFor(x => x.Address).NotEmpty().MaximumLength(300);
         RuleFor(x => x.Email).MaximumLength(150);
         RuleFor(x => x.Phone).MaximumLength(50);
