+++ docker/kernel/project/ViajesAltairis.Application/Features/Admin/Hotels/Validators/UpdateHotelValidator.cs
@@ -10,7 +10,7 @@ public class UpdateHotelValidator : AbstractValidator<UpdateHotelCommand>
         RuleFor(x => x.Id).GreaterThan(0);
         RuleFor(x => x.CityId).GreaterThan(0);
         RuleFor(x => x.Name).NotEmpty().MaximumLength(200);
-        RuleFor(x => x.Stars).InclusiveBetween((byte)1, (byte)5);
+        RuleFor(x => x.Stars).InclusiveBetween(1, 5);
         RuleFor(x => x.Address).NotEmpty().MaximumLength(300);
         RuleFor(x => x.Email).MaximumLength(150);
         RuleFor(x => x.Phone).MaximumLength(50);
